Separando la funcion de displaySearchResult a una .js
